Translation pt-br
Hello Developers, this is my translation of the site into Brazilian Portuguese. I am happy to be able to help this wonderful community. I apologize for having lengthened some sentences in the translation, this is due to our grammar rules (which are several by the way), but I hope I have helped the community in some way, and if this translation is approved, you can count on me in future translations.